Add files via upload
This notebook is part of a Coursera project that showcases the use of Python for geospatial data analysis and interactive map visualization. Using libraries like folium and geography, the project plots a launch site on a map, calculates distances to the nearest city, railway, and highway, and visualizes these connections with lines. This project highlights key skills in Python mapping and geospatial analysis